:truck: Refs #168 - Rename files
